small cleanup
In the cleanup PR, we added the `Data` and `Actions` type, but we
already had a `Actions` type so had to rename it to something. Chose
`Command` but this is now inconsistent with the rest of the codebase.

Instead, let's revert that change and use these shorthands:

- `Data` -> `_Data`
- `Actions` -> `_Actions`
- `Commands` -> `Actions`
- `CommandTypes` -> `ActionTypes`

The `_` prefix is a little bit strange, but it is a private type and not
exposed so fine for now.