Allow `<button>` to be in nested components in `<PopoverButton>` (#2715)
* Tweak `dom()` helper

The helper can currently return a component instance when it should only ever return a DOM element. So, we fix the implementation to return null if it’s not an `Element` _and_ adjust the types such that if a `ComponentPublicInstance` is passed we change the return type to `Element`.

* Specialize DOM helper to HTML elements

Technically it could be an SVG element but much of Headless UI assumes HTML elements all over. So we’ll adjust the types to assume HTMLElement instead.

* Allow `dom()` helper to return any `Node` type

It doesn’t actually always return an HTMLElement but we have behavior that relies on it returning and checking for `Comment` nodes

* Detect `<button>` nested in components inside `<PopoverButton>`

* Update changelog